+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   - Split into PollHandler, PollModel, PollOptionModel and PollAnswerModel
   - PollHandler handles API calls, sockets interaction, EventListeners etc.
   - PollModel along the nested Models handles data preparation, updates, validation and rendering
+- Moved where users data for "display results" is stored from `user.flags` to `sessionStorage`
+- Updated PollDialog from Dialog to DialogV2 
 
 ## v1.1.1
 * Added number of Answers to table on the Saved Polls App